Task 0 solved

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -1,8 +1,13 @@
 #include "lists.h"
 
+/**
+ * print_listint - Prints each node of a singly linked list
+ * @h: Pointer to head of singly linked list
+ * Return: Number of nodes in the singly linked list
+ */
 size_t print_listint(const listint_t *h)
 {
-	unsigned int count = 0;
+	size_t count = 0;
 
 	while (h != NULL)
 	{
